Add or remove hashes when FAQ's are opened or closed
We want to be able to link to FAQ's on the jetpack-connect pricing page and have them be open when linked to. Under the circumstance that somebody were to go to the link with a hash already in the URL, and then click a link that leads to the same FAQ, the FAQ would not be opened since no change in hash was detected. To prevent this, these changes remove the hash anytime the FAQ is closed, so if it were to be linked to again, it would then detect a hash changed and open up the FAQ. The hash's are also added to the URL when the FAQ is opened for user convenience if they were to want to send that link to somebody else with the FAQ already open